requirements.txt
# Core Dependencies
langgraph>=0.2.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-ollama>=0.1.0

# PDF Processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0
pymupdf>=1.23.0

# Web Scraping & APIs
requests>=2.31.0
beautifulsoup4>=4.12.0
arxiv>=2.1.0
aiohttp>=3.9.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
matplotlib>=3.7.0
httpx[http2]>=0.24.0
lxml>=4.9.2
python-dotenv>=1.0.0
seaborn>=0.12.0

# GUI Framework
streamlit>=1.28.0
plotly>=5.17.0

# Async Support
asyncio-throttle>=1.0.2
aiofiles>=23.2.1

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0

# Development Tools
black>=23.9.0
flake8>=6.1.0
mypy>=1.6.0

# Environment Management
python-dotenv>=1.0.0

# Logging
structlog>=23.2.0

# Model Utilities
transformers>=4.35.0
torch>=2.1.0